Migrate ignore tests to separate files
Will make it easier to add new tests
