+++ tests/inputs-ignore/multiline-block-ignore-1.lua
@@ -1,0 +1,5 @@
+local foo     =      bar
+-- stylua: ignore start
+local bar   =     baz
+-- stylua: ignore end
+local bar   =     baz
