+++ tests/inputs-ignore/multiline-block-ignore-2.lua
@@ -1,0 +1,6 @@
+local foo     =      bar
+-- stylua: ignore start
+local bar   =     baz
+local bar   =     baz
+-- stylua: ignore end
+local bar   =     baz
