+++ tests/inputs-ignore/multiline-block-ignore-no-ending.lua
@@ -1,0 +1,5 @@
+local foo     =      bar
+-- stylua: ignore start
+local bar   =     baz
+local bar   =     baz
+local bar   =     baz
