+++ tests/inputs-ignore/multiline-block-ignore-no-starting.lua
@@ -1,0 +1,5 @@
+local foo     =      bar
+local bar   =     baz
+local bar   =     baz
+-- stylua: ignore end
+local bar   =     baz
