+++ tests/inputs-ignore/multiline-block-ignore-scope-no-ending.lua
@@ -1,0 +1,7 @@
+local foo     =      bar
+do
+    -- stylua: ignore start
+    local bar   =     baz
+    local bar   =     baz
+end
+local bar   =     baz
