+++ tests/inputs-ignore/multiline-block-ignore-scope.lua
@@ -1,0 +1,8 @@
+local foo     =      bar
+do
+    -- stylua: ignore start
+    local bar   =     baz
+    -- stylua: ignore end
+    local bar   =     baz
+end
+local bar   =     baz
