+++ tests/inputs-ignore/singleline-ignore-1.lua
@@ -1,0 +1,3 @@
+local foo     =      bar
+-- stylua: ignore
+local bar   =     baz
