+++ tests/inputs-ignore/singleline-ignore-2.lua
@@ -1,0 +1,4 @@
+local foo     =      bar
+-- stylua: ignore
+local bar   =     baz
+local bar   =     baz
