+++ tests/inputs-ignore/singleline-ignore-last-stmt.lua
@@ -1,0 +1,2 @@
+-- stylua: ignore
+return      "hi"
