+++ tests/inputs-ignore/singleline-ignore-stmt-block.lua
@@ -1,0 +1,5 @@
+local   x     = 1
+-- stylua: ignore
+function foo   ()
+    return    x +    1
+end
